requirements.txt
hydra-core==1.2.0
pre-commit==2.14.1
black==22.6.0
flake8==3.8.4
python-json-logger==2.0.2
opencv-contrib-python==4.5.4.58
torch==1.12.0
torchvision==0.13.0
matplotlib==3.5.2
omegaconf==2.2.2
tensorboard==2.9.1
natsort==7.1.1
tqdm==4.64.0
scipy==1.8.1
imageio==2.21.1
pygifsicle==1.0.5